Pack wiki into YAML for lame FSs
I made the assumption that POSIX semantics (== case sensitive, only
disallowed chars are / and null) would work everywhere I needed.

But, of course, that was a silly assumption.

So now the new idea is to just build this on a good POSIX FS and package
that up into a single file.  The code is nicer this way, anyway.